requirements.txt
Flask>=0.12
twilio~=6.0.0
word2number
pyodbc
ngrok
openai
uuid
ast
firebase_admin
datetime
